Added lombok, fixed couple of inconsistencies

+++ src/main/java/distributed/systems/das/presentation/BattleFieldViewer.java
@@ -100,7 +100,7 @@ public class BattleFieldViewer extends JPanel implements Runnable {
 
 				/* Draw healthbar */
 				doubleBufferGraphics.setColor(Color.GREEN);
-				filler = (double)yRatio * u.getHitPoints() / (double)u.getMaxHitPoints();
+				filler = yRatio * u.getHitPoints() / (double)u.getMaxHitPoints();
 				doubleBufferGraphics.fillRect((int)(x + 0.75 * xRatio), (int)(y + 1 + yRatio - filler), (int)xRatio / 4, (int)(filler));
 
 				/* Draw the identifier */
